+++ src/Repository/Collmex/ApiClient/BaseRequest/BaseRequestAwareInterface.php
@@ -1,0 +1,11 @@
+<?php
+
+namespace App\Repository\Collmex\ApiClient\BaseRequest;
+
+interface BaseRequestAwareInterface
+{
+    /**
+     * @return BaseRequestInterface|null
+     */
+    public function getBaseRequest(): ?BaseRequestInterface;
+}