+++ src/Repository/Collmex/Type/TypeIdentifier.php
@@ -1,0 +1,10 @@
+<?php
+
+namespace App\Repository\Collmex\Type;
+
+enum TypeIdentifier: string
+{
+    case CMXDLV = 'CMXDLV';
+
+    case CMXPRD = 'CMXPRD';
+}