+++ CHANGELOG.md
@@ -4,7 +4,16 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+See [ROADMAP.md](ROADMAP.md) for roadmap and release planning.
+
 ## [Unreleased]
+
+## [0.2.0] - 2022-03-19
+### Added
+- GraphQL schema and resolver for Collmex type "Delivery" ("Lieferung")
+- ROADMAP.md
+- DateTimeConverter for Collmex date string
+- Development notest for new types in README.md
 
 ## [0.1.0] - 2022-03-01
 ### Added
@@ -12,6 +21,6 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Symfony Flex App
 - overblog/graphql-bundle configuration
 - mjaschen/collmex API client
-- GraphQL schema and resolver for Collmex type "Product"
+- GraphQL schema and resolver for Collmex type "Product" ("Produkt")
 - CHANGELOG.md
 - README.md